requirements.txt
numpy
pillow
scipy
scikit-learn
matplotlib
tqdm
